extract run support bundle routes
